Week1_PDF_Summarizer/requirements.txt
# Core dependencies
PyMuPDF
openai
tiktoken
rich
python-dotenv

# Development dependencies
pytest